% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_baseline.R
\name{BEE.calc.baseline}
\alias{BEE.calc.baseline}
\title{Calculate baseline value for each date and pixel.}
\usage{
BEE.calc.baseline(
  YourSpatraster,
  start_date = NULL,
  end_date = NULL,
  threshold,
  quantile_value = NULL,
  time_window = 5,
  smooth_window = 10
)
}
\arguments{
\item{YourSpatraster}{the SpatRaster that contains the values to be used for
calculating the baseline (your time serie of reference). It must have a
date for each layer and no duplicated date.}

\item{start_date}{the date to use as the first day of your reference period.
It must be in the format YYYY-MM-DD.}

\item{end_date}{the date to use as the last day of your reference period.
It must be in the format YYYY-MM-DD.}

\item{threshold}{tells whether you want to use a percentile or the mean of
the observed value as threshold. Accepted arguments are "qt" to specify you
want to compute a percentile of the observed values in the reference
timeframe, or "mean" if you want to use the mean of the observed values in
the timeframe provided as threshold.
If you want to use a fixed value as a threshold, you can skip this step and
directly use BEE.calc.binarize().}

\item{quantile_value}{indicates the desired percentile value. This must be
between 0 and 1.}

\item{time_window}{number of days on either side of day 'd' that are used to
calculate the threshold value on 'd' day. For instance if time_window = 5,
the value will be calculated using data from five days before day d, day d
itself, and five days after day d, from all years between start_date and
end_date. (The example assume that you have one layer per day.)}

\item{smooth_window}{number of days on either side of day 'd' that are used
to compute a mean value assigned to the 'd' day. For example, if
smooth_window is set to 10 days, the final value on day d will be the mean
of the baseline values calculated for days d - 10 to d + 10
(eleven values).}
}
\value{
A SpatRaster with one day of the year per layer (366 layers), having
the same extent, pixel resolution and crs than the provided SpatRaster.
}
\description{
Calculate baseline value for each date and pixel.
}
