% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_corrections.R
\name{BEE.calc.corrections}
\alias{BEE.calc.corrections}
\title{Compute the impact of the selected definition of extreme events on the number
of days of extreme event}
\usage{
BEE.calc.corrections(Events_corrected)
}
\arguments{
\item{dataframe}{The second element of the output of BEE.calc.true_event
(a list of data.table containing information about the value of each pixel
before and after applying criteria to distinguish anormal days from extreme
events).}
}
\value{
A text summarising the modifications induced.
}
\description{
Computes the number of modifications between the raw binerised
Spatraster and the Spartraster produced using BEE.calc.true_event. This
provides an indication of the differences between the number of days above
the threshold/baseline and the number of days that are considered to be
extreme events, according to the criteria set in BEE.calc.true_event.
}
\examples{
TO BE ADD


}
