% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_escape.R
\name{BEE.calc.escape}
\alias{BEE.calc.escape}
\title{Computes metrics relative to the distance to the border of the extreme event
(in surface, 2D)}
\usage{
BEE.calc.escape(
  extreme_events_output,
  start_date = NULL,
  end_date = NULL,
  pixel,
  only_days_EE = FALSE,
  group_by_event = FALSE
)
}
\arguments{
\item{extreme_events_output}{:
Is the ouput of function BEE.id.extreme_events().}

\item{start_date}{:
Defines the beginning of the timeframe during which you want to analyse
distance to escape. If no dates are provided, computation will be done
on all days_provided.}

\item{end_date}{:
Defines the endding of the timeframe during which you want to analyse
distance to escape. If no dates are provided, computation will be done
on all days_provided}

\item{pixel}{:
Is a vector with x,y coordinates or a df of x,y coordinates or "all" if you
want to compute distance metrics for all pixels in the raster provided
through extreme_events_output.}

\item{only_days_EE}{:
= 'TRUE' if you want to perform calculations using all the days.
Use account = 'FALSE' when you want to calculate only for the days when your
reference point is experiencing an extreme event.
By default, only_days_EE is set on FALSE}

\item{group_by_event}{:
Allows to specify if you want your result by extreme event or for the whole
time periode specified. "group_by_event" = TRUE : compute one mean distance
per event. "group_by_event" = FALSE : mean distance will be based on
distances from all days in the time frame (belonging to an extreme event or
not), including all days belonging to this event event if they are a '0'
day, TRUE : compute one mean distance accounting only for days that belong
to an EE.}
}
\value{
A list with one dataframe per pixel. Please make sure to call element of the
list by their names and not by their positions.
}
\description{
This function calculate the median distance, mean distance and standard
deviation of distance to escape form an extreme event. Two methods are used:
First is the "strait" distance, measured 'as the crow files'this method
doesn't avoid the NA pixels and consider the earth as a sphere for distance
computation.
The second method, "bio", also account for the sphericity of the earth and
compute distances over trajectory that avoid NA pixel. To do so the
trajectory is built step by step which induce more imprecision than the
"strait" method.
If the difference between the two distances is small than the spatraster
resolution, it is very likely that the difference is due to the lower
precision of the "bio" methods. If the differences is close or superior to
the spatraster resolution, it is very likely that it is because the "bio"
distance take in account the extra distance caused by going round an
obstacle (NA pixel).
(see the package vignette for more detail).
The function also provided the azimut between the pixel to flee and the
closest refuge pixel according each method.
}
\note{
All distance metrics are limited by the size of the Spatraster your
are providing. See function is not developped to work on 4D data (time + 3D).
}
\examples{
# Load data for example: library(BioExtremeEvent)
# reconstruct the output of BEE.id.extreme_event() from data saved in local:
file_name_1 <- system.file(file.path("extdata",
                                     "binarized_corrected_spatraster.tiff"),
                                     package = "BioExtremeEvent")
extreme_event_spatraster <- terra::rast(file_name_1)
file_name_2 <- system.file(file.path("extdata",
                                     "binarized_corrected_df.rds"),
                                     package = "BioExtremeEvent")
extreme_event_list <- readRDS(file_name_2)

extreme_event <- list(extreme_event_spatraster,extreme_event_list)

# create a dataframe with the positions of the studdied sites:
GPS <- data.frame(x = c(3.7659),
                 y = c(43.4287))

#' # Get distance, azimut and pixel coordinates to escape:
library(BioExtremeEvent)

escape <- BioExtremeEvent::BEE.calc.escape(
 extreme_events_output = extreme_event,
 start_date = "2024-05-01",
 end_date = "2024-11-30",
 pixel= GPS,
 only_days_EE = FALSE,
 group_by_event = FALSE
)

}
