% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_data_summarise.R
\name{BEE.data.summarise}
\alias{BEE.data.summarise}
\title{Merge the outputs of the metrics functions.}
\usage{
BEE.data.summarise(
  data = data,
  variable = variable,
  summarise_by = "extreme_event"
)
}
\arguments{
\item{data}{the dataset conatainning the column you want to process (e.g. the
ouput of BEE.data.merge())}

\item{variable}{the name of the column for wich you want to compute mean,
median etc.}

\item{summarise_by}{takes the followings options:
\itemize{
\item "extreme_event" for each metric, the following will be computed for each
extreme event: mean, median, variance, minimum and maximum, and the dates of
the maximum and minimum values.
\item a numeric value indicating the time step for which you want to compute the
mean, median, variance, minimum and maximum values, as well as the dates of
the maximum and minimum values.
}

#'@return}
}
\description{
To merge the daily outputs of at least two datasets of the
following fonctions from BEE package : bee.calc.metrics_point() ;
bee.calc.metrics_morpho() ; bee.calc.escape(). This function can also
summarise metrics over different time periods: extreme events, weak periods,
monthly periods and yearly periods.
}
\examples{
# TO BE ADDED

}
