% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_merge_summarize.R
\name{BEE.merge_summarize}
\alias{BEE.merge_summarize}
\title{Merge the outputs of the metrics functions and summarize them trought time.}
\usage{
BEE.merge_summarize(
  data_metrics_point = NULL,
  data_metrics_morpho = NULL,
  data_escape = NULL,
  summarize_by,
  crs
)
}
\arguments{
\item{data_metrics_point}{the output of the BEE.calc.metrics_point computed
using the argument group_by_event = FALSE.}

\item{data_metrics_morpho}{the output of the BEE.calc.metrics_morpho}

\item{data_escape}{the output of the BEE.calc.escape computed
using the argument group_by_event = FALSE.}

\item{summarize_by}{takes the followings options, into "" :
\itemize{
\item "extreme_event" for each metrics, a mean, median, variance, min and max will
be computed for each extrem event
\item "day" metrics are not summarized through time, the function keeps a daily
resolution and just merge the datasets.
\item "week" for each metrics, a mean, median, variance, min and max will
be computed weekly. Note that this imply to provide daily output
(BEE.calc.escape(only_days_EE = FALSE))
\item "two_weeks" for each metrics, a mean, median, variance, min and max will
be computed every two weeks. Note that this imply to provide daily output
(BEE.calc.escape(only_days_EE = FALSE))
\item "month" for each metrics, a mean, median, variance, min and max will
be computed monthly. Note that this imply to provide daily output
(BEE.calc.escape(only_days_EE = FALSE))
\item "year" for each metrics, a mean, median, variance, min and max will
be computed yearly. Note that this imply to provide daily output
(BEE.calc.escape(only_days_EE = FALSE))
}}

\item{crs}{a METRICS crs that suits the studdied area}
}
\value{
A dataframe with the metrics of all the datasets provided in column,
if you haven't choose a daily resolution, a mean, median, variance, minimum
and maximum will be computed for each variables over the choosen time step.
}
\description{
To merge the daily outputs of at least two datasets of the
following fonctions from BEE package : bee.calc.metrics_point() ;
bee.calc.metrics_morpho() ; bee.calc.escape(). This function can also
summarise metrics over different time periods: extreme events, weak periods,
monthly periods and yearly periods.
}
\examples{
# TO BE ADDED
}
