% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_plot_categories.R
\name{BEE.plot.categories}
\alias{BEE.plot.categories}
\title{Plot event categories}
\usage{
BEE.plot.categories(
  metric_point_df,
  color_theme = "red",
  start_date = NULL,
  end_date = NULL,
  ...
)
}
\arguments{
\item{metric_point_df:}{The output of \emph{BEE.calc.metric_point()} for a given location. Please note
that the output of BEE.calc.metric_point() is a list. You only need to
provide a dataframe from that list, not the whole list.}

\item{color_theme:}{A vector of four colour codes of your choice (the first one is for category
1 and the last one is for category 4), or 'red' to use an automatic palette
of red shades, or 'blue' for a colour-shaded palette.}

\item{start_date:}{= NULL by default or the date at which you want to start the plot. It must
be in the same format than metric_point_df$date.}

\item{end_date:}{= NULL by default or the date at which you want to stop the plot. It must be
in the same format than metric_point_df$date.}

\item{...:}{Customising the graph is possible by adding any general ggplot2 argument.}
}
\value{
A 'ggplot2' with the date on the x-axis, the studied parameter value on the
y-axis, and the extreme events coloured according to their category.
}
\description{
Use \emph{BEE.calc.metric_point()} to plot extreme event categorie over the all
time period or over a subset of that period.
}
\examples{
 # Load data:
file_name_1 <- system.file(file.path("extdata", "metrics_points_day.rds"),
                                  package = "BioExtremeEvent")
metrics_points_day <- readRDS(file_name_1)
# Get data set for one GPS point
metrics_points_day <- metrics_points_day[[1]] # first GPS point
BEE.plot.categories(metric_point_df = metrics_points_day,
                    color_theme = "red",
                    start_date = NULL,
                    end_date = NULL)

}
