% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_plot_cum_ano_season.R
\name{BEE.plot.cumulative_anomaly}
\alias{BEE.plot.cumulative_anomaly}
\title{Cumulative anomaly against months of the years}
\usage{
BEE.plot.cumulative_anomaly(
  metric_point_df,
  start_date = NULL,
  end_date = NULL,
  ...
)
}
\arguments{
\item{metric_point_df:}{The output of \emph{BEE.calc.metric_point()} for a given location. Please note
that the output of BEE.calc.metric_point() is a list. You only need to
provide a dataframe from that list, not the whole list.}

\item{color_theme:}{A vector of four colour codes of your choice (the first one is for category
1 and the last one is for category 4), or 'red' to use an automatic palette
of red shades, or 'blue' for a colour-shaded palette.}

\item{start_date:}{= NULL by default or the date at which you want to start the plot. It must
be in the same format than metric_point_df$date.}

\item{end_date:}{= NULL by default or the date at which you want to stop the plot. It must be
in the same format than metric_point_df$date.}

\item{...:}{Customising the graph is possible by adding any general ggplot2 argument.}
}
\value{
A 'ggplot2' with the months on the x-axis, the cumulative anomaly per
extreme event of the studied parameter value on the y-axis.
}
\description{
Allows to see at which time of the year a place is explosed to the bigest
anomalies.
}
\examples{
# Load data:
file_name_1 <- system.file(file.path("extdata", "metrics_points_day.rds"),
                                  package = "BioExtremeEvent")
metrics_points_day <- readRDS(file_name_1)
# Get data set for one GPS point
metrics_points_day <- metrics_points_day[[1]] # first GPS point
BEE.plot.cumulative_anomaly(metric_point_df = metrics_points_day,
                    start_date = NULL,
                    end_date = NULL)

}
