% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioExtremeEvent-package.R
\docType{package}
\name{BioExtremeEvent-package}
\alias{BioExtremeEvent}
\alias{BioExtremeEvent-package}
\title{BioExtremeEvent: BioExtremeEvent}
\description{
The goal of this project is to bring a new view for the analysis of the marine heat wave and more precisely bring a morphological view.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VicoMarbec/BioExtremeEvent}
  \item Report bugs at \url{https://github.com/VicoMarbec/BioExtremeEvent/issues}
}

}
\author{
\strong{Maintainer}: Victoria Delannoy \email{victoria.delannoy@protonmail.com} [copyright holder]

}
\keyword{internal}
