% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_remove.R
\name{moyenne}
\alias{moyenne}
\title{Compute the arithmetic mean}
\usage{
moyenne(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A \code{numeric} representing the arithmetic mean of \code{x}.
}
\description{
This function computes the arithmetic mean of a numeric variable.
}
\examples{
x <- 1:10
moyenne(x)
}
