% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_celsius.R
\name{BEE.calc.celsius}
\alias{BEE.calc.celsius}
\title{Convert temperature data to Celsius}
\usage{
BEE.calc.celsius(YourSpatRaster)
}
\arguments{
\item{YourSpatRaster}{"YourSpatRaster" is the SpatRaster containing temperature data, for which you
want to check the units and convert the values to Celsius.}
}
\value{
YourSpatRaster
Your SpatRaster with values corrected to celsius and 'unit' metadata updated
to 'Celsisus'.
}
\description{
Function that detects the unit of measurement of a pixel value in a
SpatRaster (Celsius, Kelvin or Fahrenheit) and converts it to the
international unit 'Celsius' if necessary.
}
\examples{

### Load the example dataset in R environement :
file_name <- system.file(file.path("extdata", "copernicus_example_data.nc"),
                                  package = "BioExtremeEvent")
copernicus_data <- terra::rast(file_name)

### Values before applying functions :
head(na.omit(terra::values(x = copernicus_data[[1]],
                   dataframe=TRUE))) # first pixels are on land so their value is NA

### Apply function :
copernicus_data_celsius <- BioExtremeEvent::BEE.calc.celsius(copernicus_data)

### Values after applying functions :
head(na.omit(terra::values(x = copernicus_data_celsius[[1]],
                  dataframe=TRUE))) # first pixels are on land so their value is NA

### Checking the unit of the spatraster :
# Before using BEE.calc.celsius :
terra::units(copernicus_data[[1]])
# After using BEE.calc.celsius :
terra::units(copernicus_data_celsius[[1]])

}
