% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_celsius.R
\name{BEE.calc.celsius}
\alias{BEE.calc.celsius}
\title{Convert temperature data to Celsius}
\usage{
BEE.calc.celsius(YourSpatRaster)
}
\arguments{
\item{YourSpatRaster}{"YourSpatRaster" is the SpatRaster containing temperature data, for which you
want to check the units and convert the values to Celsius.}
}
\value{
YourSpatRaster
Your SpatRaster with values corrected to celsius and 'unit' metadata updated
to 'Celsisus'.
}
\description{
Function that detects the unit of measurement of a pixel value in a
SpatRaster (Celsius, Kelvin or Fahrenheit) and converts it to the
international unit 'Celsius' if necessary.
}
\examples{
#If your data are in kelvins (K), BEE.calc.celscius(YourSpatRaster) returns 
#the spatraster with cell values converted in celsius (<c2><b0>C) using :
#"former value - 273.15 = new value".
#A pixel with a value of 295.15 K becomes a pixel with a value of 
#22 <c2><b0>C.
}
