% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_corrections.R
\name{BEE.calc.corrections}
\alias{BEE.calc.corrections}
\title{Compute the impact of the selected definition of extreme events on the number
of days of extreme event}
\usage{
BEE.calc.corrections(Events_corrected)
}
\arguments{
\item{Events_corrected}{The second element of the BEE.calc.true_event ouput,
which is a list of data.table containing information
about the value of each pixel before and after definition criteria are
applied to distinguish isolated extreme days from extreme events).}
}
\value{
A text summarising the modifications induced.
}
\description{
It computes the number of modifications between the raw
binerised Spatraster and the Spartraster produced using the
BEE.calc.true_event.
This indicates of the differences in the number of days above the
threshold/baseline and the number of days that are considered to be
extreme events, according to the criteria set in BEE.calc.true_event.
}
\examples{
# to be added

}
