% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_metrics_morpho.R
\name{BEE.calc.metrics_morpho}
\alias{BEE.calc.metrics_morpho}
\title{Compute metrics spatial features of the extreme event (only in surface, 2D)}
\usage{
BEE.calc.metrics_morpho(
  Corrected_rasters,
  start_date = NULL,
  end_date = NULL,
  per_pix = FALSE,
  crs = NULL
)
}
\arguments{
\item{Corrected_rasters:}{The SpatRaster you want to analyse (first output of
\emph{BEE.calc.true_event()}).}

\item{start_date:}{Allows to perform the analysis on a specific  time_frame, this allows to
save computation time.}

\item{end_date:}{Allows to perform the analysis on a specific  time_frame, this allows to
save computation time.}

\item{per_pix:}{Use TRUE if you want a list with one dt per pixel as an output.}

\item{crs:}{To get accurate length and area, data in longitude latitude must
be converted into meters data that take in account earth sphericity, the
conversion depend on studdied region, thus you must specify the one the most
appropriated for your data. See help here to find a suitable EPSG you can
check on https://epsg.io/ .}
}
\description{
This function give the morphological metrics for an extreme event,
Corrected_rasters is the SpatRaster you want to analyze (it needs to be
binarise with the mhw function)
}
\note{
BEE.calc.metrics_morpho is not designed to work on 4D data
(time+spatial 3D).
}
