% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_metrics_morpho.R
\name{BEE.calc.metrics_morpho}
\alias{BEE.calc.metrics_morpho}
\title{Compute metrics spatial features of the extreme event (only in surface, 2D)}
\usage{
BEE.calc.metrics_morpho(
  Corrected_rasters,
  start_date = NULL,
  end_date = NULL,
  per_pix = FALSE
)
}
\arguments{
\item{start_date}{and @param end_date allows to perform the analysis on a
specific  time_frame, this allows to save computation time.}

\item{per_pix}{use TRUE if you want a list with one dt per pixel as an output}

\item{Corrected_raster}{the SpatRaster you want to analyse}
}
\description{
This function give the morphological metrics for an extreme
event, Corrected_raster is the SpatRaster you want to analyze (it needs to be
binarise with the mhw function)
}
\note{
Your SpatRaster need to be binarised with the mhw function before you
put it in this function
BEE.calc.metrics_morpho is not designed to work on 4D data (time+spatial 3D).
}
