% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_metrics_point.R
\name{BEE.calc.metrics_point}
\alias{BEE.calc.metrics_point}
\title{Compute metrics through time for specific locations.}
\usage{
BEE.calc.metrics_point(
  Events_corrected,
  YourSpatraster,
  GPS,
  start_date = NULL,
  end_date = NULL,
  time_lapse_vector = NULL,
  baseline_qt = NULL,
  baseline_mean = baseline_mean,
  group_by_event = FALSE
)
}
\arguments{
\item{Events_corrected}{:
Is the list of data tables produced by:
bee_calc_true_event (the second element of the output). For each pixel, it
contains a data.table with dates in the rows and a column indicating whether
it is a day belonging to a heatwave (1) or not (0).}

\item{YourSpatraster}{:
Is the spatraster with the values of the studded parameter, through time and
space.}

\item{GPS}{:
Is a data frame containing the positions for which you want to compute
metrics. It must contain columns labelled 'x' and 'y', which should contain
longitudes and latitudes, respectively.}

\item{start_date}{:
First day on which you want to start computing metrics.}

\item{end_date}{:
Last day on which you want to start computing metrics.}

\item{time_lapse_vector}{:
A vector of time laps on which to compute mean evolution rate and variance
of the studdied parameter.}

\item{baseline_qt}{:
Spatraster of the 90th percentile baseline (or the 10th percentile
baseline).}

\item{baseline_mean}{:
Spatraster of the mean value baseline.}

\item{group_by_event}{:
Whether you want an output summarise by extreme event or not. If not, you
just get daily metrics.}
}
\value{
A list of dataframe (one per GPS point), each dataframe contains
informations on the date of the extrem events, mean, median, max and min
values, peak day, onset-rate, off-set rate, mean anomaly, maximum category
\emph{etc.} Categories are defined in Hobday \emph{et al.} 2018.
}
\description{
Compute metrics through time for specific locations.
}
\note{
\emph{BEE.calc.metrics_point()} is not designed to work on 4D data
(time + spatial 3D).
}
\examples{
# Prepare function arguments:
file_name_1 <- system.file(file.path("extdata",
                                     "copernicus_data_celsius.tiff"),
                                     package = "BioExtremeEvent")
copernicus_data_celsius <- terra::rast(file_name_1)
file_name_2 <- system.file(file.path("extdata",
                                     "binarized_corrected_df.rds"),
                                     package = "BioExtremeEvent")
binarized_corrected_df <- readRDS(file_name_2)
file_name_3 <- system.file(file.path("extdata",
                                     "baseline_qt90_smth_15.tiff"),
                                     package = "BioExtremeEvent")
baseline_qt90_smth_15 <- terra::rast(file_name_3)
file_name_4 <- system.file(file.path("extdata",
                                     "baseline_mean_smth_15.tiff"),
                                     package = "BioExtremeEvent")
baseline_mean_smth_15 <- terra::rast(file_name_4)
GPS <- data.frame(x = c(3.6883,4.4268), # Sète, Saintes-Marie-de-la-Mer
                  y = c(43.3786,43.4279))

# Get daily value (required for BEE.merge() and BEE.summarise()):
metrics_points_day <- BEE.calc.metrics_point(
 Events_corrected = binarized_corrected_df,
 YourSpatraster = copernicus_data_celsius,
 GPS = GPS,
 start_date = NULL,
 end_date = NULL,
 time_lapse_vector = NULL,
# number of time unit on which to compute the warming rates and cooling rates,
#  NULL means it will not be computated
 baseline_qt = baseline_qt90_smth_15,
 baseline_mean = baseline_mean_smth_15,
 group_by_event = FALSE
)

# Get mean, min, max, sd and median per event (extreme event and btw extreme
# events):
metrics_points_ee <- BEE.calc.metrics_point(
 Events_corrected = binarized_corrected_df,
 YourSpatraster = copernicus_data_celsius,
 GPS = GPS,
 start_date = NULL,
 end_date = NULL,
 time_lapse_vector = NULL,
# number of time unit on which to compute the warming rates and cooling rates,
#  NULL means it will not be computated
 baseline_qt = baseline_qt90_smth_15,
 baseline_mean = baseline_mean_smth_15,
 group_by_event = TRUE
)

}
