% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_calc_true_event.R
\name{BEE.calc.true_event}
\alias{BEE.calc.true_event}
\title{Identify the extreme event}
\usage{
BEE.calc.true_event(binarized_EE, n, d = NULL, nbis = NULL, w = NULL, p = NULL)
}
\arguments{
\item{binarized_EE}{A list of Spatraster built using the BEE.calc.binarize
function.}

\item{n}{The minimum number of consecutive days above the baseline/threshold
required for a series of "1" to be considered as an extreme event (≥).}

\item{d}{The maximum number of days below the baseline/thershold for
considering two series of values above the threshold/baseline but separated
by some days below threshold/baseline as a single extreme event (maximum
distance in days between the two events to be merged, ≤).}

\item{nbis}{NOT AVAILBALE YET The minimum number of days in each series of
1 to allow merging if there are d days or fewer between the to series.
Example :
if you set 'nbis' to 3, short series of 1 (n) can be merged with long
series of 1 (≥n) if the two series are separated by no more than d days and
there are at least nbis days between them (with at least one of the series
having to be longer than n).
Default: NULL}

\item{w}{NOT AVAILBALE YET Window w between to event in which you
allow a certain proportion p of days (within the window) to be bellow the
threshold, windows is bordered by series of consecutive days above
threshold. w and p replace d. If w is an odd number,
the minimum number of days above threshold within w will be the value
rounded to the superior round number. For instance if w = 7 and p = 0.3
a window of 7 days between to event of n consecutive days will not
create two distinct event if at least 4 days are above threshold.
( w\emph{p = 7}0.5 = 3.5 rounded to 4).
Example : n = 5; w = 10; p = 0.5 -> w * p = 5 . Default: NULL}

\item{p}{NOT AVAILBALE YET, see "w" explanations. Default: NULL}
}
\value{
Returns a list with 2 elements, first one is called
"stacked_rasters_corrected". This is a Spatraster containing binarised values
corrected according the definition used in the function. The second element
is called "Event_corrected". This is a list with one data.table per pixel.
Each data.table as a date per rows (and as many date as there are layers in
YourSpatraster). The columns "Original_value" gives the value of the raw
binarisation, before appling the definition. "Cleanned_value" gives for
each dates the values after correction. "ID" is an ID of the event,
built using the pixel ID, first day of the event and last day of the
event. "Nb_days" indicates the duration (in unit of time) of the event.
"date" indicates the date corresponding to the row.
}
\description{
Identify the extreme event according to the set of constraints.
}
\details{
w and p settings are not available yet. BEE.calc.true_event is not
designed to work on 4D data (time + spatial 3D).
}
\examples{
# This function apply a filter to withdraw isolated events. For instance,
# is you consider that an event that last only 2 days should not be accounted
# for in future analysis, you can correct those value by 0. Here are the 
# different filter you can apply:
#
# Low complexity: 
# - Minimum number of consecutive day above threshold to 
#  consider that there is an extreme event: n  (>=)
#  
# -   Maximum number of days bellow thershold to consider two series of 
# value above threshold as a single extreme event (distance max in days btw 
# 2 events to merge them): d (<=)
#   
# Examples: n >= 5  d>= 2
#       0 0 0 1 1 1 1 ** 1 ** 0 0 0 = 1 event, 
#   but 0 0 0 1 1 1 1 0 0 0 = no event
#       0 0 0 1 1 1 1 1 0 0 1 1 1 0 0 0 = 1 event of 5 days, 
#   but 0 0 0 1 1 1 1 1 0 0 1 1 1 **1 1 ** 0 0 0 = 1 event of 12 days,
#   but 0 0 0 1 1 1 1 1 0 0 ** 0 ** 1 1 1 1 1 0 0 0 = 2 instincts 
# event of 5 days each.
#
# Medium complexity (Hobday et al. 2016, 2018):
#- Minimum number of days distant from d days or less to an event
#  of n days : nbis .
#  
# Example: n = 5, d = 2, nbis = 3
# This is for the case you want to consider 
#     0 0 0 1 1 1 1 1 0 0 1 1 ** 1 ** 0 0 0 as one event of 10 days 
# but 0 0 0 1 1 1 1 1 0 0 1 1 0 0 0 as one event of 5 days. 
# In the first case, if there are not five consecutive days above threshold
# right before or after (d = n still fixed to 5).
# 
#  High complexity : NOT AVAILABLE YET
# - Unstable window * w * between to event in which you allow a certain 
#  proportion p of days of w to be bellow threshold if this windows is 
# bordered by series of consecutive days above threshold on each size. 
# w  and  p  replace d. If w is an odd number, the minimum number of days 
# above threshold within w will be the value rounded to the superior round
#  number. For instance if w = 7 and p = 0.3 a window of 7 days between to
#  event of * n * consecutive days will not create two distinct event if at
#  least 4 days are above threshold.                                                                                                                             w * \ * * p * = 7 \ * 0.5 = 3.5 rounded to 4)
#
# Examples : n = 5
# w = 10
# p = 0.5  ->  w*p = 5 
# case that lead to a single big event of 15 days:
# 0 0 0 0 0 1 1 1 1 1 0 1 1 1 0 0 0 1 1 0 1 1 1 1 1 0 0 0 0 0
# 0 0 0 0 0 1 1 1 1 1 0 0 1 1 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 
# (you can also use low complexity n = 5 and d = 3)
# 0 0 0 0 0 1 1 1 1 1 0 1 0 1 0 1 0 1 0 1 1 1 1 1 1 0 0 0 0 0
# case that lead to 2 distinct events:
# 0 0 0 0 0 1 1 1 1 1 0 1 1 0 0 1 0 1 0 0 1 1 1 1 1 0 0 0 0 0
# 0 0 0 0 0 1 1 1 1 1 0 0 0 1 1 1 1 0 0 0 1 1 1 1 1 0 0 0 0 0
# 0 0 0 0 0 1 1 1 1 1 0 1 1 1 1 0 0 0 0 0 1 1 1 1 1 0 0 0 0 0 (
# if you want to detect 1 event of 10 days then a distinct event of 5 days
# instead of 2 events of 5 days only you can use : 
# low complexity with n = 4 and d = 1)
}
