% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_data_load_Copernicus.R
\name{BEE.data.load_copernicus}
\alias{BEE.data.load_copernicus}
\title{Download spataraster from Copernicus to your local memory.}
\usage{
BEE.data.load_copernicus(
  username,
  password,
  dataset_id,
  dataset_version,
  variables,
  minimum_longitude,
  maximum_longitude,
  minimum_latitude,
  maximum_latitude,
  start_datetime,
  end_datetime,
  coordinates_selection_method = "strict-inside",
  disable_progress_bar = FALSE,
  output_directory = here::here("Data")
)
}
\arguments{
\item{username}{is your Copernicus account username, you can create an
account here :
https://data.marine.copernicus.eu/register?redirect=\%2Fproducts\%3Fdisc\%3Dnone}

\item{password}{is your Copernicus account password.}

\item{dataset_id}{is available on the corpernicus page of your product, in
"Data access" rubric, under the temporal resolution name. It is usually
start by cmems_ ...}

\item{dataset_version}{it the version of the dataset, you can found it in
the "Product user manual" or as subtitle in the Marine Data Store, it is
usually in capital letters.}

\item{variables}{is the variable you are interested in, you can only call
one variable at a time. Variables name are visible in Marine Data Store :
Product : Data access : From (bellow subset) : use the name in grey before
square bracket}

\item{minimum_longitude}{You can use the "Draw on map" Copernicus tool to
select an area of interest in Marine Data Store : Product : Data access
: Form (bellow subset) and copy the values using "N" (north) as the maximum
latitude and "W" (west) as the minimum longitude if working in the northwest
hemisphere.}

\item{maximum_longitude}{maximum longitude at which data will be extracted.}

\item{minimum_latitude}{minimum latitude at which data will be extracted.}

\item{maximum_latitude}{maximum latitude at which data will be extracted.}

\item{start_datetime}{First day wanted in the downloaded spatraster. Please
make sure that the date is available on copernicus.}

\item{end_datetime}{Last day wanted in the downloaded spatraster. Please
make sure that the date is available on copernicus.}

\item{coordinates_selection_method}{= stric-inside" by default, extraction
method, to see more info, check the python API help to get moer detail
https://help.marine.copernicus.eu/en/articles/7949409-copernicus-marine-
toolbox-introduction}

\item{disable_progress_bar}{allows to pot a progress bar of the downloading,
by default it is set on 'FALSE'.}

\item{output_directory}{where you want the data to be save, by default it
goes to the "Data" file next to your script if there is one.}
}
\value{
A SpatRaster within the timeframe provided and the spatial frame
given.
}
\description{
#' @description This function is a wrapper from the CopernicusMarine
function from pyhton. A pyhton setup must be present in your computer, thus
this function is testing if some pyhton is present and if not it offer to
download it from R.
}
\details{
You can select your downloading conditions in
https://data.marine.copernicus.eu/products : Product : Data access :
Form and then copy and past the argument of the copernicusmarine.subset
function in https://data.marine.copernicus.eu/products > Product >
Data access : Form : Automate : Pyhton API and use them as this function
arguments.
}
