% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_data_merge.R
\name{BEE.data.merge}
\alias{BEE.data.merge}
\title{Merge the outputs of the metrics functions.}
\usage{
BEE.data.merge(
  data_metrics_point = NULL,
  data_metrics_morpho = NULL,
  data_escape = NULL,
  crs
)
}
\arguments{
\item{data_metrics_point}{the output of the BEE.calc.metrics_point computed
using the argument group_by_event = FALSE.}

\item{data_metrics_morpho}{the output of the BEE.calc.metrics_morpho}

\item{data_escape}{the output of the BEE.calc.escape computed
using the argument group_by_event = FALSE.}

\item{crs}{a METRICS crs that suits the studdied area}
}
\value{
A dataframe with the metrics of all the datasets provided in column,
if you haven't choose a daily resolution, a mean, median, variance, minimum
and maximum will be computed for each variables over the choosen time step.
}
\description{
To merge the daily outputs of at least two datasets of the
following fonctions from BEE package : bee.calc.metrics_point() ;
bee.calc.metrics_morpho() ; bee.calc.escape(). This function can also
summarise metrics over different time periods: extreme events, weak periods,
monthly periods and yearly periods.
}
\examples{
# TO BE ADDED

}
