% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_data_summarise.R
\name{BEE.data.summarise}
\alias{BEE.data.summarise}
\title{Summarise the outputs of the 'merge' function.}
\usage{
BEE.data.summarise(
  data = data,
  variable = variable,
  summarise_by = "extreme_event"
)
}
\arguments{
\item{data:}{The dataset containing the column you want to process (e.g. the ouput of
\emph{BEE.data.merge()}, as a list of data frame).}

\item{variable:}{The name of the column for which you want to compute mean, median etc. over
a time window.}

\item{summarise_by:}{Takes the following options:
\itemize{
\item "extreme_event" for each metric, the following will be computed for each
extreme event: mean, median, variance or sd, minimum and maximum, sum, and
the dates of the maximum and minimum values.
\item a numeric value indicating the time step for which you want to compute the
mean, median, etc.
}}
}
\value{
A list of dataframes, with one dataframe per pixel, with the dates in the
rows and the values summarised.
}
\description{
To summarise one of the daily outputs of the following fonctions from BEE
package : \emph{bee.calc.metrics_point()} ; \emph{bee.calc.metrics_morpho()} ;
\emph{bee.calc.escape()} ; \emph{bee.data.merge()}. Metrics values can be summarised
per extreme event or over a sliding time window.
}
\details{
In addition to the mean value, the function provides the median, standard
deviation, minimum and maximum values, sum, date of minimum and maximum
values (within the time window), and date of minimum value (also within the
time window).
}
\examples{
# TO BE ADDED

}
